backend/requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.17
djangorestframework-simplejwt>=5.5
django-cors-headers>=4.9
mysqlclient>=2.2
